+++ nugu-agent/src/main/java/com/skt/nugu/sdk/agent/display/AbstractDisplayAgent.kt
@@ -15,7 +15,7 @@
  */
 package com.skt.nugu.sdk.agent.display
 
-import com.skt.nugu.sdk.core.interfaces.capability.AbstractCapabilityAgent
+import com.skt.nugu.sdk.agent.AbstractCapabilityAgent
 import com.skt.nugu.sdk.core.interfaces.context.ContextManagerInterface
 import com.skt.nugu.sdk.core.interfaces.context.PlayStackManagerInterface
 import com.skt.nugu.sdk.core.interfaces.focus.FocusManagerInterface
