+++ nugu-agent/src/main/java/com/skt/nugu/sdk/agent/microphone/AbstractMicrophoneAgent.kt
@@ -15,7 +15,7 @@
  */
 package com.skt.nugu.sdk.agent.microphone
 
-import com.skt.nugu.sdk.core.interfaces.capability.AbstractCapabilityAgent
+import com.skt.nugu.sdk.agent.AbstractCapabilityAgent
 import com.skt.nugu.sdk.core.interfaces.context.ContextManagerInterface
 import com.skt.nugu.sdk.core.interfaces.message.MessageSender
